requirements.txt
python-fasthtml
uvicorn>=0.29
python-dotenv~=1.0.1
Markdown~=3.7
litellm~=1.55.2
pyyaml~=6.0.2